[Win] Deadlock in WTF::Thread::didExit() while WebKitNetworkProcess.exe is exiting https://bugs.webkit.org/show_bug.cgi?id=210955
Reviewed by Don Olmstead.

The thread_local object of the main thread is destructed after
Windows terminates other threads. If the terminated thread was
holding a mutex, trying to lock the mutex causes deadlock.

* wtf/win/ThreadingWin.cpp:
(WTF::Thread::ThreadHolder::~ThreadHolder): Do nothing if the
thread is the main thread.

Canonical link: https://commits.webkit.org/224101@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@260911 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WTF/wtf/win/ThreadingWin.cpp
@@ -267,6 +267,12 @@ void Thread::establishPlatformSpecificHandle(HANDLE handle, ThreadIdentifier thr
 struct Thread::ThreadHolder {
     ~ThreadHolder()
     {
+        // The thread_local object of the main thread is destructed
+        // after Windows terminates other threads. If the terminated
+        // thread was holding a mutex, trying to lock the mutex causes
+        // deadlock.
+        if (isMainThread())
+            return;
         if (thread) {
             thread->specificStorage().destroySlots();
             thread->didExit();
